Tweaking
I have pretty much finished all of the website, and at this point I am just tweaking bits and bobs to match the design pixel perfect as much as I possibly can.

I have used a heading that was not mentioned in the style guide which is h6. I have used it instead of h5 as it matches the design more accurately than h5 does even at specified pixel size.

I think that the elements after adjusting are pixel perfect and if not they are almost pixel perfect.